1.서브페이지 수정 2.헤더 수정 3.FAQ-관리자페이지로 이동(임시)

+++ src/main/java/com/spike/dto/LoanDTO.java
@@ -34,7 +34,7 @@ public class LoanDTO {
 			generator = "loan_seq_generator"
 			)
 	
-	private Integer loanId;
+	private Long loanId;
 	
 	@ManyToOne
 	private UserDTO owner;
